move showCommonness to a component
